# fix latency measurement distribution