Fixes #46.

Previous version of the code called RecordCorrectedValues
on Histogram that resulted in compensation of all missed
request points with latency interpolation and then executed
all queued requests according to the schedule.

This patch fixes use of HdrHistogram by recording corrected
latencies per every executed request manually.

Further improvements:

- don't return latency from test functor
- allow switching measurements modes between Operation, Intended (corrected latency according to a schedule) and Both
- add Start to limiter to postpone first request into the future
- rename rate limiter