@pveentjer 2021-01-12
I'll have a look. But by all means, I'm not a Go programmer :) So I can only have a look at the logic.