@psarna 2021-01-13
@sitano could you explain to me how to observe the difference and how did you verify the solution? Preferably with some reproducible steps.
I tried to check the following scenario:
1. Spawn a single-node Scylla cluster
2. Start a scylla-bench workload
3. Pause the Scylla node for a couple of seconds with `kill -STOP`
4. Restore the Scylla node with `kill -CONT`
5. Observe what happened to scylla-bench during (and after) the time when Scylla was not responsive

Meanwhile, I observed the output both with and without the patch, but I didn't see any difference. Is it expected? How to deterministically verify that this patch computes latencies differently?