@sitano 2021-01-14
@psarna with my pleasure! I will start with some context to make it more clear. The problem itself is described in the issue https://github.com/scylladb/scylla-bench/issues/46 by Peter and relates to the Coordinated Omission problem.

## Coordinated Omission

The coordination omission (CO) is a class of mistakes in the implementation of Open Class load generators that split between:

- load generation
- latency measurement

Omitting all the details of what happens otherwise, a correct implementation must do the following things:

- generate load according to a schedule because its generation muscle is limited by the parallelism factor (number of worker threads) - that's why we are having Wait() function in the RateLimiter that waits for the time point when the next request is expected   to be executed.
- handle case when it can't keep up with a schedule. this happens when the latency of a request grows or spikes due to overload. in the worst case a worker can't output required amount of requests in a second time unit. at the best case generated workload slightly decline from the schedule and returns back to it.

Now when a worker is out of a schedule it can do 2 things:

- compensate missed requests into the latency distribution and proceed from the next point in the schedule. thats for `HdrHistrogram.RecordIntendedValue` is. It saves a measured latency + adds latencies for all missed requests into the histogram. 
- queue requests that were not executed in time and then execute them immediately ASAP. when it's doing that it must save correct latency. what does it mean to be correct? it means that the final latency distribution must reflect that the system was overloaded. for an open class queueing system what is the request latency? it is time from the moment it was queued and executed. when it was queued for execution? since its intended point in the schedule! so the `correct latency is = start_time - intended_time + latency = end_time - intended_time`.

the latter is what I am fixing.

## so how it can be tested?

you must create a situation of system overload!

why? because in that situation workers would not be able to keep up with the intended load they are issued to create. that will result into a situation that they will either float around the intended load schedule or around throughput equilibrium. so that they will no be able to keep up with an intended load schedule. that will result into latency correction necessity. then your old and new results will differ.

how to test it.

1. run Scylla in docker and limit its CPU power: `docker run -it scylladb/scylla --smp 1`
2. run the previous version on it to find out what max throughput 1 core is capable of having: `./scylla-bench -mode write -workload uniform --duration 1m --consistency-level one -nodes 272592d62266`. let's say its 10000 QPS.
3. now test 99%% for prev and new version:

    $ ./scylla-bench -mode write -workload uniform --duration 1m --consistency-level one -nodes 272592d62266 --max-rate 9000
   and
    $ ./scylla-bench -mode write -workload uniform --duration 1m --consistency-level one -nodes 272592d62266 --max-rate 11000

You will see the difference in 99%%. For new version it will be more than 100ms. For old version it will be about 2-4ms.

## Example:

My core can handle about 12500 ops/core.

### old version

```
./scylla-bench -mode write -workload uniform --duration 1m --consistency-level one -nodes 40aec34fbbac -max-rate 13000
Configuration
Mode:			 write
Workload:		 uniform
Timeout:		 5s
Consistency level:	 one
Partition count:	 10000
Clustering rows:	 100
Clustering row size:	 Fixed(4)
Rows per request:	 1
Page size:		 1000
Concurrency:		 16
Connections:		 4
Maximum rate:		 13000 op/s
Client compression:	 true

 time   ops/s  rows/s errors max    99.9th 99th   95th   90th   median mean   
   1s   11736   11736      0 5.4ms  4.3ms  2.6ms  2ms    1.5ms  1.3ms  1.4ms  
   2s   14268   14268      0 17ms   14ms   5.3ms  1.6ms  1.4ms  918µs  1.1ms  
   3s   13000   13000      0 2.2ms  1.9ms  1.6ms  1.2ms  1.1ms  426µs  532µs  
   4s   13005   13005      0 2ms    1.8ms  1.5ms  1.2ms  1.1ms  426µs  527µs  
   5s   13000   13000      0 2.4ms  1.9ms  1.5ms  1.2ms  1.1ms  426µs  519µs  
   6s   12995   12995      0 2.3ms  1.8ms  1.6ms  1.2ms  1ms    426µs  514µs  
   7s   13007   13007      0 2ms    1.8ms  1.5ms  1.2ms  1.1ms  426µs  529µs  
   8s   13004   13004      0 2.4ms  1.9ms  1.6ms  1.2ms  1.1ms  426µs  521µs  
   9s   12992   12992      0 22ms   22ms   12ms   1.6ms  1.4ms  983µs  1.2ms  ^C
...
```
### new version

#### overload

```
./scylla-bench -mode write -workload uniform --duration 1m --consistency-level one -nodes 40aec34fbbac -max-rate 13000
Configuration
Mode:			 write
Workload:		 uniform
Timeout:		 5s
Consistency level:	 one
Partition count:	 10000
Clustering rows:	 100
Clustering row size:	 Fixed(4)
Rows per request:	 1
Page size:		 1000
Concurrency:		 16
Connections:		 4
Maximum rate:		 13000 op/s
Client compression:	 true

 time   ops/s  rows/s errors max    99.9th 99th   95th   90th   median mean   
   1s   11913   11913      0 88ms   86ms   85ms   81ms   78ms   46ms   44ms   
   2s   11791   11791      0 182ms  180ms  177ms  173ms  169ms  134ms  133ms  
   3s   12583   12583      0 226ms  224ms  220ms  215ms  211ms  194ms  192ms  
   4s   14662   14662      0 227ms  225ms  217ms  209ms  199ms  146ms  148ms  
   5s   13757   13757      0 146ms  144ms  139ms  132ms  126ms  94ms   74ms   
   6s   11813   11813      0 121ms  120ms  116ms  110ms  105ms  67ms   67ms   
   7s   11771   11771      0 217ms  217ms  213ms  207ms  204ms  167ms  167ms  
   8s   15713   15713      0 233ms  231ms  227ms  220ms  214ms  87ms   94ms   
   9s   13008   13008      0 8ms    7.1ms  4.1ms  2.6ms  2.2ms  1.1ms  1.4ms  ^C
interrupted
```

#### non overload

```
./scylla-bench -mode write -workload uniform --duration 1m --consistency-level one -nodes 40aec34fbbac -max-rate 12000
Configuration
Mode:			 write
Workload:		 uniform
Timeout:		 5s
Consistency level:	 one
Partition count:	 10000
Clustering rows:	 100
Clustering row size:	 Fixed(4)
Rows per request:	 1
Page size:		 1000
Concurrency:		 16
Connections:		 4
Maximum rate:		 12000 op/s
Client compression:	 true

 time   ops/s  rows/s errors max    99.9th 99th   95th   90th   median mean   
   1s   11964   11964      0 10ms   8.9ms  7.2ms  6ms    5.5ms  3.1ms  3ms    
   2s   12059   12059      0 6.2ms  5.9ms  4.8ms  3.7ms  2ms    852µs  1.1ms  
   3s   12013   12013      0 2.2ms  1.9ms  1.6ms  1.3ms  1.1ms  492µs  568µs  
   4s   12002   12002      0 4ms    3.3ms  2.7ms  1.9ms  1.6ms  655µs  812µs  
   5s   12003   12003      0 4.5ms  3.9ms  2.8ms  1.9ms  1.6ms  819µs  908µs  
   6s   12013   12013      0 4.2ms  3.4ms  2.6ms  1.8ms  1.5ms  688µs  810µs  
   7s   12012   12012      0 3.5ms  2.9ms  1.9ms  1.4ms  1.2ms  524µs  642µs  
   8s   12005   12005      0 2.4ms  1.9ms  1.6ms  1.2ms  1ms    459µs  549µs  
   9s   12014   12014      0 4.1ms  3.4ms  2.7ms  1.8ms  1.4ms  557µs  720µs  
  10s   12010   12010      0 2.2ms  2ms    1.6ms  1.3ms  1ms    459µs  552µs  ^C
interrupted
```